+++ src/xhydro/indicators/__init__.py
@@ -1,0 +1,7 @@
+"""Indicators submodule."""
+
+# Import the submodules
+from . import generic, pmp
+
+# Specific top-level functions
+from .generic import *
