+++ src/xhydro/indicators/generic.py
